Initial release
Most things are working but barely tested. It doesn't yet
use information about the most common letters nor can
find words using anything but the beginning of the word. It
also doesn't test if all lines are as long. Although that may
not be a problem, it's not supported.

Everything is subject to change.

Uses some code from my sudoku-solver's development version.